+++ Practica_4/algorRec.h
@@ -1,3 +1,3 @@
-int factorialrec(int n);
-int fiborec(int n);
+double factorialrec(int n);
+double fiborec(int n);
 void hanoi(int n, char from_rod, char to_rod, char aux_rod);
